requirements.txt
astroid==2.2.5
autopep8==1.4.4
backports.csv==1.0.7
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
# Django==2.2.3
Django==2.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==4.0.0
django-simple-captcha==0.5.11
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
flake8==3.7.7
future==0.17.1
hanziconv==0.3.2
httplib2==0.9.2
idna==2.8
isort==4.3.19
jdcal==1.4.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
# psycopg2-binary==2.9.1
# ~~PyMySQL==1.0.2~~
# ~~mysqlclient~~
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pymongo==3.8.0
python-dotenv==0.19.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.4.2
urllib3==1.25.3
wrapt==1.11.1
# xadmin-django==3.0.1
xlrd==1.2.0
xlwt==1.3.0
